fix deserialization
